+++ app/src/main/kotlin/ac/mdiq/podcini/storage/model/Volume.kt
@@ -1,5 +1,6 @@
 package ac.mdiq.podcini.storage.model
 
+import ac.mdiq.podcini.storage.database.allFeeds
 import ac.mdiq.podcini.storage.database.deleteFeed
 import ac.mdiq.podcini.storage.database.getFeedList
 import ac.mdiq.podcini.storage.database.realm
@@ -148,7 +149,7 @@ fun Volume.allChildren(): List<Volume> {
 
 suspend fun deleteVolumeTree(volume: Volume) {
     Logd(TAG, "deleteVolumeTree volume: ${volume.name}")
-    val feeds_ = realm.query(Feed::class).query("volumeId == ${volume.id}").find()
+    val feeds_ = allFeeds.filter { it.volumeId == volume.id }
     for (f in feeds_) {
         val worthyEps = f.worthyEpisodes
         deleteFeed(f.id, worthyEps.isNotEmpty())